fixed wrong type

+++ src/analysis/processing/qgsalgorithmimportphotos.cpp
@@ -279,7 +279,7 @@ QVariantMap QgsImportPhotosAlgorithm::processAlgorithm( const QVariantMap &param
   outFields.append( QgsField( QStringLiteral( "directory" ), QVariant::String ) );
   outFields.append( QgsField( QStringLiteral( "altitude" ), QVariant::Double ) );
   outFields.append( QgsField( QStringLiteral( "direction" ), QVariant::Double ) );
-  outFields.append( QgsField( QStringLiteral( "orientation" ), QVariant::Double ) );
+  outFields.append( QgsField( QStringLiteral( "orientation" ), QVariant::Int ) );
   outFields.append( QgsField( QStringLiteral( "longitude" ), QVariant::String ) );
   outFields.append( QgsField( QStringLiteral( "latitude" ), QVariant::String ) );
   outFields.append( QgsField( QStringLiteral( "timestamp" ), QVariant::DateTime ) );
